Revert "Add font and color customization UI for RTF rendering with compact la…"
This reverts commit 699b4cefd10076a6c65f9db921aaa8b334329f78.
