+++ QLStephenSwift/QLStephenSwift/QLStephenSwiftApp.swift
@@ -32,8 +32,9 @@ struct QLStephenSwiftApp: App {
     var body: some Scene {
         Window("QLStephenSwift", id: "main") {
             ContentView()
-                .frame(minWidth: 500, maxWidth: 500, minHeight: 600, maxHeight: 800)
+                .frame(width: 500, height: 660)
         }
+        .windowResizability(.contentSize)
         .commands {
             CommandGroup(replacing: .newItem) { }
         }
